finished image uploading, final tests pending
