+++ server/controllers/account/edit.controller.ts
@@ -2,6 +2,7 @@ import { FastifyRequest, FastifyReply } from 'fastify';
 import fsSync from 'fs';
 import { promises as fs } from 'fs';
 import path  from 'path';
+import uploadAvatar from '../utils/file.controller';
 const USERMANAGEMENT_URL: string = process.env.USERMANAGEMENT_URL || "http://user_container:3000";
 
 export async function editUsername(request: FastifyRequest, reply: FastifyReply) {
@@ -62,50 +63,12 @@ export async function editAvatar(request: FastifyRequest, reply: FastifyReply) {
 	try {
 		const { username } = request.params as { username: string };
 
-		// get file
+		//// get file
 		const file = await request.file();
 		if (!file) {
-		  throw { code: 406, message: 'No content' };
+			throw { code: 406, message: 'No content' };
 		}
-		const extension = file.filename.split('.').pop();
-		
-		// get upload directory
-		const buffer = await file.toBuffer();
-		const wrkdir = process.cwd();
-		const uploadDir = path.join(wrkdir, 'public/images');
-
-		// ensure uploads directory exists
-		if (!fsSync.existsSync(uploadDir)) {
-			fsSync.mkdirSync(uploadDir, { recursive: true });
-		}
-
-		// upload
-		const filePath = path.join(uploadDir, file.filename);
-		if (!fsSync.existsSync(filePath)) {
-			fsSync.writeFile(filePath, buffer, (err) => {
-				if (err) {
-					throw { code: 500, message: "Erroring uploading file" };
-				}});
-			}
-		
-		const userAvatar = (username + "_avatar") as string;
-		const rand = (Math.random() * 10).toString(10).substr(0, 5);
-		const filename = (username + "_avatar_E" + rand + "." + extension) as string;
-
-		// delete existing avatar
-		const files = fsSync.readdirSync(uploadDir);
-		files.forEach(f => {
-			if (f.split('_E')[0] == userAvatar)
-			{
-				// find existing with username_avatar name
-				const deleteMe = path.join(uploadDir, f);
-				fsSync.unlinkSync(deleteMe); // delete file
-			}
-		});
-		
-		const oldPath = filePath;
-		const newPath = path.join(uploadDir, filename);
-		await fs.rename(oldPath, newPath);
+		const filename = await uploadAvatar(username, file);
 
 		// update dashboard info with new user info + return new view
 		const resEdit = await fetch(`${USERMANAGEMENT_URL}/editAvatar/${username}`, {
